DRYD-1062: Upgrade fcart vocabularies to 7.1.
The purpose of this is to add the annotationType vocabulary, which was erroneously missing from fcart config. In order to (hopefully) avoid conflicts when upgrading to 7.1, this updates the file to what is currently in master for 7.1. As a consequence, some other vocabularies will also be added that won't be used until 7.1.